moved check license_text_exists to own file
