+++ src/ros_license_toolkit/main.py
@@ -25,11 +25,12 @@ import timeit
 from typing import Optional, Sequence
 
 from ros_license_toolkit.checks import (LicenseTagExistsCheck,
-                                        LicenseTagIsInSpdxListCheck,
-                                        LicenseTextExistsCheck, Status)
+                                        LicenseTagIsInSpdxListCheck, Status)
 from ros_license_toolkit.license_file_referenced_check import \
     LicenseFilesReferencedCheck
 from ros_license_toolkit.license_in_code_check import LicensesInCodeCheck
+from ros_license_toolkit.license_text_exists_check import \
+    LicenseTextExistsCheck
 from ros_license_toolkit.package import get_packages_in_path
 from ros_license_toolkit.ui_elements import (FAILURE_STR, SUCCESS_STR,
                                              WARNING_STR, Verbosity, major_sep,
